# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Autozapi",
    "description": "Serviço para comunicação com o WhatsApp",
    "version": "1.0.0"
  },
  "host": "localhost:3333",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/{applicationId}/{sessionId}/qrcode": {
      "get": {
        "description": "Obtem o qrcode de autenticação",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "description": "'base64' | 'urlCode'",
            "schema": {
              "type": "string",
              "example": "base64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{applicationId}/{sessionId}/status": {
      "get": {
        "description": "Obtem o status da conexção",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{applicationId}/{sessionId}/send/text": {
      "post": {
        "description": "Envia uma mensagem de texto para um determinado número",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "chatId": {
                  "type": "string",
                  "example": "553799999999@g.us"
                },
                "text": {
                  "type": "string",
                  "example": "foo"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{applicationId}/{sessionId}/send/audio": {
      "post": {
        "description": "Envia uma mensagem de audio no formato \"ptt\", para um determinado número",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "chatId": {
                  "type": "string",
                  "example": "553799999999@g.us"
                },
                "base64": {
                  "type": "string",
                  "example": "..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{applicationId}/{sessionId}/send/image": {
      "post": {
        "description": "Envia uma imagem para um determinado número",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "chatId": {
                  "type": "string",
                  "example": "553799999999@g.us"
                },
                "base64": {
                  "type": "string",
                  "example": "..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{applicationId}/{sessionId}/logout": {
      "post": {
        "description": "Realiza o logout do dipositivo",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "chatId": {
                  "example": "any"
                },
                "base64": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}